+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
-- Updated embedded FrozenLib to resolve a crash with Lithium.
-- Updated Sodium compatability.
+- Removed Particle Rain compat.